pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pbp-batch"
version = "0.0.3"
description = "Running mbari-pbp library for full deployments"
authors = [
    { name = "Xavier Mouy", email = "xavier.mouy@whoi.edu" }
]
license = { text = "BSD-3-Clause" }
requires-python = ">=3.11, <3.12" 
dependencies = [
    "mbari-pbp==1.8.74",
    "prefect",
    "prefect-dask",
    "pydantic>=2.0",
    "pyyaml",
    "pandas",
    "matplotlib",
    "bokeh>=2.4.2,!=3.0.*",
    "soundfile",
    "panel",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
[tool.setuptools]
packages = ["pbp_batch"]

[project.scripts]
pbp-batch = "pbp_batch.cli:main"
